requirements.txt
gunicorn==20.1.0
flask==2.1.3
jinja2==3.0.1
psycopg2-binary==2.9.3
Flask-SQLAlchemy==2.5.1
GitPython==3.1.27
libsast==1.5.2